modules/module4-llm-providers/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
openai-agents
pytest
pytest-asyncio
google-generativeai